NETMIKO/venv_netmiko_textfsm/conf/requirements.txt
python-dotenv
paramiko
netmiko
rich
textfsm

